Check auth inside services

+++ module/Conferences/src/Conferences/Service/ConferenceService.php
@@ -3,6 +3,7 @@
 namespace Conferences\Service;
 
 use Conferences\Entity\Conference;
+use BjyAuthorize\Exception\UnAuthorizedException;
 
 /**
  * Handles interaction with events (IE conferences)
@@ -14,11 +15,15 @@ class ConferenceService {
 
     private $entityManager;
     
+    private $authorize;
+    
     private $conferenceRepository;
     
-    public function __construct($entityManager) {
+    public function __construct($entityManager, $authorize) {
 
         $this->entityManager = $entityManager;
+        
+        $this->authorize = $authorize;
 
         $this->conferenceRepository = $this->entityManager->getRepository('Conferences\Entity\Conference');
         
@@ -44,6 +49,11 @@ class ConferenceService {
     
     public function upsertConference(\Application\Entity\Systemuser $I_loggedUser, Conference $conference) {
         
+        // check if logged user has enough privileges to edit this conf
+        if (!$this->authorize->isAllowed($conference, 'edit')) {
+            throw new UnAuthorizedException();
+        }
+        
         $conference->setOwner($I_loggedUser);
         
         $this->entityManager->persist($conference);
@@ -54,6 +64,11 @@ class ConferenceService {
     }
     
     public function removeConference(Conference $conference) {
+        
+        // check if logged user has enough privileges to edit this conf
+        if (!$this->authorize->isAllowed($conference, 'edit')) {
+            throw new UnAuthorizedException();
+        }
         
         $this->entityManager->remove($conference);
         $this->entityManager->flush();
